Using pipes, inbuilt and custom made
Added duration pipe to show duration as string and not a number. Also created pipes for filtering. Used OnChanges to use the filter only when filterBy property changes in session-list.component.ts

Used a visibleSessions array to filter out those sessions from the sessions array that match the level user searched for. For Example if user clicked on the all button then i am filtering out all sessions that have an 'all' attribute on them.  Next we will create a pipe for sorting data since angular 2 unlike angular 1 does not provide us with a sort pipe due to performance issues in creating a generic sorting pipe.
We will use PipeTransform and Pipe decorator to do so